Firemaking, questing and cache.
What has been done today:

Added the basics for firemaking (tinderbox on item in inventory) - Still need to add lighting on ground item, candle lighting and more.

Added a super super basic questing system, it is still incomplete but is basic
What needs to be added to it:
- Quest complete check
- Config updating
- Button id handling.

Added object definitions to be loaded from cache. - Still need to fully replace the XML based one but this is good for now.

That shall all be done in good time though.